Update index.html for Bingo Event
The title of the page has been updated to reflect the Bingo event of Hack Kosice 2024. A link to a favicon has been added and the tasks in the script have been updated with descriptions of different activities for the event. Other minor layout and styling adjustments have been made to improve functionality and user experience.